[SEDONA-708] Separate Catalog implementation into AbstractCatalog and Catalog
